requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.3
azure-common==1.1.24
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
beautifulsoup4==4.8.2
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
dj-database-url==0.5.0
Django==2.2.6
django-admin-honeypot==1.1.0
django-celery==3.3.1
django-crispy-forms==1.8.1
django-google-analytics-app==4.4.0
django-heroku==0.3.1
django-social-share==1.3.2
django-storages==1.8
django-widget-tweaks==1.4.5
gunicorn==19.7.1
idna==2.8
image==1.5.27
kombu==3.0.37
Pillow==7.0.0
psycopg2==2.8.4
pycparser==2.19
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
structlog==19.2.0
urllib3==1.25.7
whitenoise==3.3.1
django-bootstrap4==1.1.1
